// sdk/src/generated/models/ArticleType.ts
/**
 * Stack Overflow for Teams API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Article Type
*/
export enum ArticleType {
    KnowledgeArticle = 'knowledgeArticle',
    Announcement = 'announcement',
    Policy = 'policy',
    HowToGuide = 'howToGuide'
}